+++ server/src/main/java/com/example/tyfserver/auth/util/AuthorizationExtractor.java
@@ -1,8 +1,9 @@
 package com.example.tyfserver.auth.util;
 
 import com.example.tyfserver.auth.exception.AuthorizationHeaderNotFoundException;
-import java.util.Enumeration;
+
 import javax.servlet.http.HttpServletRequest;
+import java.util.Enumeration;
 
 public class AuthorizationExtractor {
