+++ server/src/main/java/com/example/tyfserver/donation/service/DonationService.java
@@ -2,7 +2,6 @@ package com.example.tyfserver.donation.service;
 
 import com.example.tyfserver.donation.domain.Donation;
 import com.example.tyfserver.donation.dto.DonationMessageRequest;
-import com.example.tyfserver.donation.dto.DonationRequest;
 import com.example.tyfserver.donation.dto.DonationResponse;
 import com.example.tyfserver.donation.exception.DonationNotFoundException;
 import com.example.tyfserver.donation.repository.DonationRepository;
