+++ server/src/main/java/com/example/tyfserver/member/controller/MemberController.java
@@ -2,7 +2,6 @@ package com.example.tyfserver.member.controller;
 
 import com.example.tyfserver.auth.dto.LoginMember;
 import com.example.tyfserver.auth.service.AuthenticationService;
-import com.example.tyfserver.common.util.S3Connector;
 import com.example.tyfserver.member.dto.*;
 import com.example.tyfserver.member.exception.NicknameValidationRequestException;
 import com.example.tyfserver.member.exception.PageNameValidationRequestException;
