+++ server/src/main/java/com/example/tyfserver/member/dto/TokenValidationRequest.java
@@ -3,10 +3,6 @@ package com.example.tyfserver.member.dto;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
