+++ server/src/main/java/com/example/tyfserver/member/repository/MemberQueryRepository.java
@@ -3,7 +3,6 @@ package com.example.tyfserver.member.repository;
 import com.example.tyfserver.member.dto.CurationsResponse;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface MemberQueryRepository {
     List<CurationsResponse> findCurations();
