+++ server/src/main/java/com/example/tyfserver/payment/domain/PaymentServiceConnector.java
@@ -1,6 +1,5 @@
 package com.example.tyfserver.payment.domain;
 
-import com.example.tyfserver.payment.domain.PaymentInfo;
 import com.example.tyfserver.payment.dto.PaymentRequest;
 
 public interface PaymentServiceConnector {
