+++ server/src/main/resources/logback-spring.xml
@@ -26,7 +26,7 @@
     </appender>
 
     <root level="INFO">
-        <appender-ref ref="Console" />
+        <appender-ref ref="Console"/>
         <appender-ref ref="ASYNC_SLACK"/>
     </root>
 </configuration>
