+++ server/src/main/resources/static/iamport/index.html
@@ -2,8 +2,8 @@
 <html lang="en">
 
 <head>
-  <meta charset="UTF-8">
-  <title>Title</title>
+    <meta charset="UTF-8">
+    <title>Title</title>
 </head>
 
 <body>
@@ -17,51 +17,51 @@
 <button onclick="requestPay()">결제하기</button>
 
 <script>
-  IMP.init('imp61348931');
+    IMP.init('imp61348931');
 
-  function requestPay() {
-    // IMP.request_pay(param, callback) 호출
-    IMP.request_pay({ // param
-      pg: "kakaopay",
-      pay_method: "card",
-      merchant_uid: "ORDjoyykim-10",
-      name: "주문1",
-      amount: 1000,
-      buyer_email: "gildong@gmail.com",
-      buyer_name: "홍길동",
-      buyer_tel: "010-4242-4242",
-      buyer_addr: "서울특별시 강남구 신사동",
-      buyer_postcode: "01181"
-    }, function (rsp) { // callback
-      if (rsp.success) { // 결제 성공 시: 결제 승인 또는 가상계좌 발급에 성공한 경우
-        // jQuery로 HTTP 요청
-        console.log(rsp);
-        jQuery.ajax({
-          url: "http://localhost:8080/payments/complete", // 가맹점 서버
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          data: {
-            "impUid": rsp.imp_uid,
-            "merchantUid": rsp.merchant_uid
-          }
-        }).done(function (data) {
-          // 가맹점 서버 결제 API 성공시 로직
-          switch (data.status) {
-            case "vbankIssued":
-              // 가상계좌 발급 시 로직
-              console.log("가상계좌 발급 완료");
-              break;
-            case "success":
-              // 결제 성공 시 로직
-              console.log("결제 성공");
-              break;
-          }
-        })
-      } else {
-        alert("결제에 실패하였습니다. 에러 내용: " + rsp.error_msg);
-      }
-    });
-  }
+    function requestPay() {
+        // IMP.request_pay(param, callback) 호출
+        IMP.request_pay({ // param
+            pg: "kakaopay",
+            pay_method: "card",
+            merchant_uid: "ORDjoyykim-10",
+            name: "주문1",
+            amount: 1000,
+            buyer_email: "gildong@gmail.com",
+            buyer_name: "홍길동",
+            buyer_tel: "010-4242-4242",
+            buyer_addr: "서울특별시 강남구 신사동",
+            buyer_postcode: "01181"
+        }, function (rsp) { // callback
+            if (rsp.success) { // 결제 성공 시: 결제 승인 또는 가상계좌 발급에 성공한 경우
+                // jQuery로 HTTP 요청
+                console.log(rsp);
+                jQuery.ajax({
+                    url: "http://localhost:8080/payments/complete", // 가맹점 서버
+                    method: "POST",
+                    headers: {"Content-Type": "application/json"},
+                    data: {
+                        "impUid": rsp.imp_uid,
+                        "merchantUid": rsp.merchant_uid
+                    }
+                }).done(function (data) {
+                    // 가맹점 서버 결제 API 성공시 로직
+                    switch (data.status) {
+                        case "vbankIssued":
+                            // 가상계좌 발급 시 로직
+                            console.log("가상계좌 발급 완료");
+                            break;
+                        case "success":
+                            // 결제 성공 시 로직
+                            console.log("결제 성공");
+                            break;
+                    }
+                })
+            } else {
+                alert("결제에 실패하였습니다. 에러 내용: " + rsp.error_msg);
+            }
+        });
+    }
 </script>
 
 </html>
